@x1ddos 2018-01-18
Thanks so much @didrocks! Will take a look at it soon.
And of course thanks @samtstern for convincing you :)

/cc @marcacohen @cassierecher in case you guys want to take a look too