@samtstern 2018-01-18
@x1ddos I told @didrocks I would take on the work of writing some tests / doing other cleanup if this PR gets merged.  As I showed you in an email this MD parser is really solid and has a bunch of features we could use.