@x1ddos 2018-01-24
Hey, sorry for a long wait. Looking at it now and will probably continue tomorrow.