@x1ddos 2018-01-24
I just want to make sure this won't introduce incompatible changes with internal usage.