@x1ddos 2018-01-30
Sorry, have an urgent change to make, to close #34. Will come back to this right after.