@x1ddos 2018-01-30
Meanwhile, @didrocks could you rebase this onto latest master?
I switched to Travis CI, to make sure it still compiles and passes all existing tests.