@didrocks 2018-01-31
And rebased! The existing tests pass locally and on Travis.