@x1ddos 2018-06-18
So, I've just ran some simple benchmarking using the following function:

```go
// testdata/codelab.md is
// https://github.com/canonical-websites/tutorials.ubuntu.com/blob/master/tutorials/server/access-remote-desktop/access-remote-desktop.md
func BenchmarkParser(b *testing.B) {
	lab, err := ioutil.ReadFile("testdata/codelab.md")
	if err != nil {
		b.Fatal(err)
	}
	r := bytes.NewReader(lab)
	b.ResetTimer()
	for i := 0; i < b.N; i++ {
	        p := Parser{}
		if _, err := p.Parse(r); err != nil {
			b.Fatal(err)
		}
	}
}
```

The results using existing parser:

```
# Old MD parser.
$ go test -bench=. -benchmem
goos: linux
goarch: amd64
pkg: github.com/googlecodelabs/tools/claat/parser/md
BenchmarkParser-4   	 1000000	      2120 ns/op	   10240 B/op	      14 allocs/op
PASS
ok  	github.com/googlecodelabs/tools/claat/parser/md	2.158s
```

The new parser in this change:

``` 
# New MD parser.
$ go test -bench=. -benchmem
goos: linux
goarch: amd64
pkg: github.com/googlecodelabs/tools/claat/parser/md
BenchmarkParser-4   	  500000	      2703 ns/op	   10920 B/op	      22 allocs/op
PASS
ok  	github.com/googlecodelabs/tools/claat/parser/md	1.396s
```

I have to admit I was expecting worse due to switching away from tokenizer-based parser, but it's not that bat at all.  

Hey @didrocks would you mind resolving conflicts? There's only one change in parse.go which adds support for `alt` and `title` image attributes. I briefly checked this change and didn't see anything similar in `func image(ds *docState) types.Node`.

I'm reviewing the code meanwhile.

It's too bad the tests were removed but @samtstern promised he would add them back. :)