@samtstern 2018-06-18
@x1ddos I went and sent a PR doing the work of fixing the merge conflicts:
https://github.com/didrocks/codelab-ubuntu-tools/pull/2

In the event that @didrocks is no longer interested in continuing this work we can use my fork of his branch.