@didrocks 2018-06-19
Rebased and nitpicks modifications done. Thanks for the review! CI passes on 1.8.x and 1.10.x.